test : Location 클래스의 변경으로 인한 테스트 코드 수정
